use strongly typed configuration; put Game behind an interface
