+++ forge-game/src/main/java/forge/game/GameSnapshot.java
@@ -47,7 +47,7 @@ public class GameSnapshot {
         }
         GameRules currentRules = origGame.getRules();
         Match newMatch = new Match(currentRules, newPlayers, origGame.getView().getTitle());
-        newGame = new Game(newPlayers, currentRules, newMatch);
+        newGame = new GameImpl(newPlayers, currentRules, newMatch);
         restore = false;
         assignGameState(origGame, newGame, includeStack);
         //System.out.println("Storing game state with timestamp of :" + origGame.getTimestamp());
