+++ forge-gui-desktop/src/test/java/forge/ai/simulation/SimulationTest.java
@@ -13,11 +13,7 @@ import forge.ai.AIOption;
 import forge.ai.LobbyPlayerAi;
 import forge.ai.simulation.GameStateEvaluator.Score;
 import forge.deck.Deck;
-import forge.game.Game;
-import forge.game.GameRules;
-import forge.game.GameStage;
-import forge.game.GameType;
-import forge.game.Match;
+import forge.game.*;
 import forge.game.card.Card;
 import forge.game.card.CardCollectionView;
 import forge.game.card.CardFactory;
@@ -44,11 +40,11 @@ public class SimulationTest {
         players.add(new RegisteredPlayer(d1).setPlayer(new LobbyPlayerAi("p1", options)));
         GameRules rules = new GameRules(GameType.Constructed);
         Match match = new Match(rules, players, "Test");
-        Game game = new Game(players, rules, match);
+        Game game = new GameImpl(players, rules, match);
         game.setAge(GameStage.Play);
-        game.EXPERIMENTAL_RESTORE_SNAPSHOT = false;
-        game.AI_TIMEOUT = FModel.getPreferences().getPrefInt(FPref.MATCH_AI_TIMEOUT);
-        game.AI_CAN_USE_TIMEOUT = true; //Only Android is restricted according to API Level
+        game.configuration().set(GameOptions.EXPERIMENTAL_RESTORE_SNAPSHOT, false);
+        game.configuration().set(GameOptions.AI_TIMEOUT, FModel.getPreferences().getPrefInt(FPref.MATCH_AI_TIMEOUT));
+        game.configuration().set(GameOptions.AI_CAN_USE_TIMEOUT, true); //Only Android is restricted according to API Level
 
         return game;
     }
