+++ forge-gui/src/main/java/forge/gamemodes/match/HostedMatch.java
@@ -153,11 +153,11 @@ public class HostedMatch {
         SoundSystem.instance.setBackgroundMusic(this.matchPlaylist == null ? MusicPlaylist.MATCH : this.matchPlaylist);
 
         game = match.createGame();
-        game.EXPERIMENTAL_RESTORE_SNAPSHOT = FModel.getPreferences().getPrefBoolean(FPref.MATCH_EXPERIMENTAL_RESTORE);
-        game.AI_TIMEOUT = FModel.getPreferences().getPrefInt(FPref.MATCH_AI_TIMEOUT);
+        game.configuration().set(GameOptions.EXPERIMENTAL_RESTORE_SNAPSHOT, FModel.getPreferences().getPrefBoolean(FPref.MATCH_EXPERIMENTAL_RESTORE));
+        game.configuration().set(GameOptions.AI_TIMEOUT, FModel.getPreferences().getPrefInt(FPref.MATCH_AI_TIMEOUT));
         // Android API 31 and above can use completeOnTimeout -> CompletableFuture:
         //https://developer.android.com/reference/java/util/concurrent/CompletableFuture#completeOnTimeout(T,%20long,%20java.util.concurrent.TimeUnit)
-        game.AI_CAN_USE_TIMEOUT = !GuiBase.isAndroid() || GuiBase.getAndroidAPILevel() > 30;
+        game.configuration().set(GameOptions.AI_CAN_USE_TIMEOUT, !GuiBase.isAndroid() || GuiBase.getAndroidAPILevel() > 30);
 
         StaticData.instance().setSourceImageForClone(FModel.getPreferences().getPrefBoolean(FPref.UI_CLONE_MODE_SOURCE));
